+++ stories/manifold-data-provision-button.stories.js
@@ -6,10 +6,12 @@ storiesOf('Provision Button [Data]', module)
   .add(
     'default',
     () => `
-      <label for="my-provision-button">Resource Name</label>
-      <input id="my-provision-button" value="my-resource">my-resource</input>
-      <manifold-data-provision-button resource-label="my-resource">
-        🚀 Provision Business plan on Prefab.cloud
-      </manifold-data-provision-button>
+      <manifold-connection>
+        <label for="my-provision-button">Resource Name</label>
+        <input id="my-provision-button" value="my-resource">my-resource</input>
+        <manifold-data-provision-button resource-label="my-resource">
+          🚀 Provision Business plan on Prefab.cloud
+        </manifold-data-provision-button>
+      </manifold-connection>
     `
   );