+++ CHANGELOG.md
@@ -5,13 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Changed the provision button so it fetches the owner ID automatically if not set.
 
-### Deprecated
-- Deprecated `resource-label` attibute on `manifold-data-product-logo`. Use `manifold-data-resource-logo` component instead.
-
+## [v0.5.0]
 ### Added
 - Added a SSO data button and the resource wrapper for ssoing into a resource's product dashboard.
 - Added a new CTA slot in the product card for displaying unique cta content.
+
+### Deprecated
+- Deprecated `resource-label` attribute on `manifold-data-product-logo`. Use `manifold-data-resource-logo` component instead.
 
 ### Fixed
 - Fixed the provision button requiring the label to be set, preventing or automatic label generation from working.